merge para el sprint 6

See merge request mayo.ortiz.adria/projecte4_entrebicis!14

+++ Backend/src/main/java/com/copernic/backend/Backend/controller/web/MatrixController.java
@@ -42,7 +42,7 @@ public class MatrixController {
             @RequestParam("tempsMaxAturat") String tempsMaxAturatStr,
             @RequestParam("tempsRecollida") String tempsRecollidaStr,
             @RequestParam("velMax") Double velMax,
-            @RequestParam("conversioSaldo") Integer conversioSaldo,
+            @RequestParam("conversioSaldo") Double conversioSaldo,
             Model model
     ) {
         try {
